Library Import Fixes
Publishing updates so this library correctly exposes the
`'net.java.dev.jna:jna:4.5.1@aar' dependency; otherwise, applications
using the library have to specify it as a dependency explicitly.

The key point is that the dependency specification for jna in
the `pom.xml` associated with e3db-client-android has the `aar` type
associated with. For example:

```
    <dependency>
      <groupId>net.java.dev.jna</groupId>
      <artifactId>jna</artifactId>
      <version>4.5.1</version>
      <type>aar</type>
      <exclusions>
        <exclusion>
          <groupId>*</groupId>
          <artifactId>*</artifactId>
        </exclusion>
      </exclusions>
    </dependency>
```