[WASH-938] Make realm info request in coroutine
- Places the request to public realm info endpoint in a coroutine
  and suspends execution of caller thread until its completion

Note that this is flagged as an innappropriate use of the execute()
method in the coroutine as it is a blocking method. Essentially,
its use makes the coroutine redundant (forgoes the usefulness of them),
but enables us to block caller thread until realmInfo is populated
by a network request made on a coroutine instead of main thread.